# docs/swagger.yaml
basePath: /
definitions:
...
  httptransport.AddTenantMemberRequest:
    properties:
      username:
        maxLength: 256
        type: string
    required:
    - username
    type: object
...
  httptransport.CreateUserRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - display_name
    - username
    type: object
...
  httptransport.DeleteUserRequest:
    properties:
      id:
        type: string
      version:
        type: integer
    required:
    - id
    - version
    type: object
...
  httptransport.GetUserRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  httptransport.PageUsersRequest:
    properties:
      created_at_from:
        type: string
      created_at_to:
        type: string
      emails:
        items:
          type: string
        maxItems: 200
        type: array
      ids:
        items:
          type: string
        maxItems: 200
        type: array
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      phones:
        items:
          type: string
        maxItems: 200
        type: array
      statuses:
        items:
          $ref: '#/definitions/identity.Status'
        maxItems: 20
        type: array
      usernames:
        items:
          type: string
        maxItems: 200
        type: array
    type: object
...
  httptransport.RemoveTenantMemberRequest:
    properties:
      id:
        type: string
      version:
        type: integer
    required:
    - id
    - version
    type: object
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
...
  httptransport.SwitchTenantRequest:
    properties:
      tenant_id:
        type: string
    required:
    - tenant_id
    type: object
  httptransport.TenantMemberIDRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  httptransport.TenantMemberPageRequest:
    properties:
      ids:
        items:
          type: string
        maxItems: 200
        type: array
      joined_from:
        type: string
      joined_to:
        type: string
      keyword:
        maxLength: 256
        type: string
      page:
        type: integer
      page_size:
        type: integer
      statuses:
        items:
          $ref: '#/definitions/tenant.Status'
        maxItems: 20
        type: array
      user_ids:
        items:
          type: string
        maxItems: 200
        type: array
      usernames:
        items:
          type: string
        maxItems: 200
        type: array
    type: object
...
  httptransport.UpdateTenantMemberStatusRequest:
    properties:
      id:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/tenant.Status'
        enum:
        - active
        - disabled
      version:
        type: integer
    required:
    - id
    - status
    - version
    type: object
...
  httptransport.UpdateUserRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      status:
        $ref: '#/definitions/identity.Status'
      version:
        type: integer
    required:
    - display_name
    - id
    - status
    - version
    type: object
...
  identity.Status:
    enum:
    - active
    - disabled
    - locked
    - closed
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusDisabled
    - StatusLocked
    - StatusClosed
  identity.User:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      status:
        $ref: '#/definitions/identity.Status'
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
      version:
        type: integer
    type: object
...
  tenant.AvailableTenant:
    properties:
      is_administrator:
        type: boolean
      joined_at:
        type: string
      membership_id:
        type: string
      tenant_code:
        type: string
      tenant_id:
        type: string
      tenant_name:
        type: string
    type: object
  tenant.ContextToken:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      tenant:
        $ref: '#/definitions/tenant.AvailableTenant'
      token_type:
        type: string
    type: object
...
  tenant.Member:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      display_name:
        type: string
      id:
        type: string
      joined_at:
        type: string
      status:
        $ref: '#/definitions/tenant.Status'
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      user_id:
        type: string
      username:
        type: string
      version:
        type: integer
    type: object
...
  tenant.Status:
    enum:
    - active
    - disabled
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusDisabled
...
info:
  contact: {}
  description: 'Production-oriented Go Web API scaffold. Application error codes:
    0 success; 10000-19999 common/input; 20000-29999 authentication/authorization;
    30000-39999 business; 50000-59999 infrastructure.'
  title: Go API Template
  version: "1.0"
paths:
...
  /api/v1/tenant-context/available:
    post:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/tenant.AvailableTenant'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: List the current user's available tenants
      tags:
      - tenant-context
  /api/v1/tenant-context/current:
    post:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/tenant.AvailableTenant'
              type: object
      security:
      - Bearer: []
      summary: Get the validated current tenant context
      tags:
      - tenant-context
  /api/v1/tenant-context/switch:
    post:
      parameters:
      - description: Tenant
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.SwitchTenantRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/tenant.ContextToken'
              type: object
      security:
      - Bearer: []
      summary: Issue an access token scoped to one tenant membership
      tags:
      - tenant-context
...
  /api/v1/tenant-members/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Member
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.AddTenantMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/tenant.Member'
              type: object
      security:
      - Bearer: []
      summary: Add a user to the current tenant
      tags:
      - tenant-members
  /api/v1/tenant-members/get:
    post:
      parameters:
      - description: Member
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.TenantMemberIDRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/tenant.Member'
              type: object
      security:
      - Bearer: []
      summary: Get a tenant member
      tags:
      - tenant-members
  /api/v1/tenant-members/page:
    post:
      parameters:
      - description: Filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.TenantMemberPageRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Page current tenant members
      tags:
      - tenant-members
  /api/v1/tenant-members/remove:
    post:
      parameters:
      - description: Member
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.RemoveTenantMemberRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Remove a tenant member
      tags:
      - tenant-members
...
  /api/v1/tenant-members/status/update:
    post:
      parameters:
      - description: Status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.UpdateTenantMemberStatusRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/tenant.Member'
              type: object
      security:
      - Bearer: []
      summary: Enable or disable a tenant member
      tags:
      - tenant-members
...
  /api/v1/users/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/identity.User'
              type: object
      security:
      - Bearer: []
      summary: Create a globally identified user
      tags:
      - users
  /api/v1/users/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: User and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Logically close and delete a user
      tags:
      - users
  /api/v1/users/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.GetUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/identity.User'
              type: object
      security:
      - Bearer: []
      summary: Get a user by ID
      tags:
      - users
  /api/v1/users/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: User filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.PageUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Search users by username, email, phone, status, and time range
      tags:
      - users
  /api/v1/users/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: User and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/identity.User'
              type: object
      security:
      - Bearer: []
      summary: Update a user using optimistic locking
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"